refactor(ui): Refactor all general components
Refactor all general components to use queries and types from the hey-api
query client (apart from `ApiError`, which will be refactored separately).
