+++ ui/src/components/charts/vulnerability-metrics.tsx
@@ -17,7 +17,6 @@
  * License-Filename: LICENSE
  */
 
-import { Vulnerability } from '@/api/requests';
 import { Cvss4RadarChart } from '@/components/charts/cvss4-radar-chart';
 import { Cvss4VectorCard } from '@/components/charts/cvss4-vector-card';
 import { Cvss23RadarChart } from '@/components/charts/cvss23-radar-chart';
@@ -28,6 +27,7 @@ import {
   findHighestCvssScore,
   findHighestEpssScore,
 } from '@/helpers/vulnerability-statistics';
+import { Vulnerability } from '@/hey-api';
 
 type VulnerabilityMetricsProps = {
   vulnerability: Vulnerability;
